modal display changes
